[mutable] Fixed incorrect total Mutable texture size reporting due to an int32 overflow. #jira UE-205298 #rnx
#tests PIE



#virtualized

[CL 31043127 by alexei lebedev in 5.4 branch]

+++ Engine/Plugins/Experimental/Mutable/Source/CustomizableObject/Internal/MuCO/LogBenchmarkUtil.h
@@ -90,7 +90,7 @@ public:
 	void FinishUpdateImage(const FString& CustomizableObjectPathName, const FString& InstancePathName, double TaskUpdateImageTime, const int64 TaskUpdateImageMemoryPeak, const int64 TaskUpdateImageRealMemoryPeak) const;
 	
 	DECLARE_BENCHMARK_STAT(NumAllocatedTextures, uint32);
-	DECLARE_BENCHMARK_STAT(TextureGPUSize, uint32);
+	DECLARE_BENCHMARK_STAT(TextureGPUSize, uint64);
 
 	DECLARE_BENCHMARK_STAT(NumAllocatedSkeletalMeshes, int32);
